+++ src/agentadaptor.h
@@ -26,7 +26,7 @@ public Q_SLOTS:
     QString RequestPinCode(const QDBusObjectPath &device, const QDBusMessage &msg);
     Q_NOREPLY void DisplayPinCode(const QDBusObjectPath &device, const QString &pincode);
     quint32 RequestPasskey(const QDBusObjectPath &device, const QDBusMessage &msg);
-    Q_NOREPLY void DisplayPasskey(const QDBusObjectPath &device, quint32 passkey, quint8 entered);
+    Q_NOREPLY void DisplayPasskey(const QDBusObjectPath &device, quint32 passkey, quint16 entered);
     void RequestConfirmation(const QDBusObjectPath &device, quint32 passkey, const QDBusMessage &msg);
     void RequestAuthorization(const QDBusObjectPath &device, const QDBusMessage &msg);
     void AuthorizeService(const QDBusObjectPath &device, const QString &uuid, const QDBusMessage &msg);
